+++ api_emulator/redfish/chassis_api.py
@@ -9,7 +9,7 @@ import g
 
 import sys, traceback
 from flask import Flask, request, make_response, render_template
-from flask.ext.restful import reqparse, Api, Resource
+from flask_restful import reqparse, Api, Resource
 
 from .templates.chassis import get_Chassis_template
 from .thermal_api import ThermalAPI, CreateThermal