+++ api_emulator/redfish/eg_resource_api.py
@@ -1,30 +1,32 @@
 # Copyright Notice:
-# Copyright 2016 Distributed Management Task Force, Inc. All rights reserved.
+# Copyright 2017 Distributed Management Task Force, Inc. All rights reserved.
 # License: BSD 3-Clause License. For full text see link: https://github.com/DMTF/Redfish-Interface-Emulator/LICENSE.md
 
-# PCIe Switches API  GET, POST
-# PCIe Switch   API  GET, PUT, PATCH, DELETE
+# Example Collection Resource and Singleton Resource
+"""
+Collection API  GET, POST
+Singleton  API  GET, PUT, PATCH, DELETE
 
+"""
 import g
 
 import sys, traceback
 from flask import Flask, request, make_response, render_template
 from flask_restful import reqparse, Api, Resource
 
-from .templates.pcie_switch import get_PCIeSwitch_template
-from .pcie_port_api import PCIePortsAPI, PCIePortAPI
-
+from .templates.eg_resource import get_EgResource_template
 
 members = []
 member_ids = []
 foo = 'false'
 INTERNAL_ERROR = 500
 
-#PCIe Switch API
-class PCIeSwitchAPI(Resource):
+#EgResource API
+class EgResourceAPI(Resource):
+    # Can't initialize the resource since the URI variables are not in the argument list.
+    # Need the variables to set the Odata.id properties
     def __init__(self):
-        print ('PCIeSwitchAPI init called')
-        print ('PCIeSwitchAPI init exit')
+        print ('EgResourceAPI init called')
 
     # HTTP GET
     def get(self,ident):
@@ -35,43 +37,45 @@ class PCIeSwitchAPI(Resource):
                     break
             config = cfg
             resp = config, 200
-        except OSError:
-            resp = error_response('Resources directory does not exist', 400)
         except Exception:
             traceback.print_exc()
             resp = INTERNAL_ERROR
         return resp
 
     # HTTP PUT
-    # - On the first call, we add the API for PCIeSwitches, because sw_id is not available in 'init'.
-    # - TODO - debug why this returns a 500 error, on the second put call
+    # - Create the resource (since URI variables are avaiable)
+    # - Update the members and members.id lists
+    # - Attach the APIs of subordinate resources (do this only once)
+    # - Finally, create an instance of the subordiante resources
     def put(self,ident):
-        print ('PCIeSwitchAPI put called')
+        print ('EgResourceAPI put called')
         try:
             global config
-            config=get_PCIeSwitch_template(g.rest_base,ident)
+            config=get_EgResource_template(g.rest_base,ident)
             members.append(config)
             member_ids.append({'@odata.id': config['@odata.id']})
             global foo
-            print ('var = ' + foo)
-            g.api.add_resource(PCIePortsAPI, '/redfish/v1/PCIeSwitches/<string:sw_id>/Ports')
-            g.api.add_resource(PCIePortAPI,  '/redfish/v1/PCIeSwitches/<string:sw_id>/Ports/<string:ident>')
+            # Attach URIs for subordiante resources
+            if  (foo == 'false'):
+                # Add APIs for subordinate resourcs
+                # g.api.add_resource(SubordinateResAPI, '/redfish/v1/EgResources/<string:ch_id>/SubordinateRes')
+                foo = 'true'
+            # Create an instance of subordinate resources
+            #cfg = CreateSubordinateRes()
+            #out = cfg.put(ident)
             resp = config, 200
-        except OSError:
-            resp = error_response('Resources directory does not exist', 400)
         except Exception:
             traceback.print_exc()
             resp = INTERNAL_ERROR
-        print ('PCIeSwitchAPI put exit')
+        print ('EgResourceAPI put exit')
         return resp
 
     # HTTP PATCH
     def patch(self, ident):
-        print ('PCIeSwitchAPI patch called')
+        print ('EgResourceAPI patch called')
         raw_dict = request.get_json(force=True)
         print (raw_dict)
         try:
-        # schema.validate(raw_dict)
             # Find the entry with the correct value for Id
             for cfg in members:
                 if (ident == cfg["Id"]):
@@ -83,8 +87,6 @@ class PCIeSwitchAPI(Resource):
                 config[key] = value
             print (config)
             resp = config, 200
-        except OSError:
-            resp = error_response('Resources directory does not exist', 400)
         except Exception:
             traceback.print_exc()
             resp = INTERNAL_ERROR
@@ -93,7 +95,7 @@ class PCIeSwitchAPI(Resource):
 
     # HTTP DELETE
     def delete(self,ident):
-        # print ('PCIeSwitchAPI delete called')
+        # print ('EgResourceAPI delete called')
         try:
             idx = 0
             for cfg in members:
@@ -103,23 +105,21 @@ class PCIeSwitchAPI(Resource):
             members.pop(idx)
             member_ids.pop(idx)
             resp = 200
-        except OSError:
-            resp = error_response('Resources directory does not exist', 400)
         except Exception:
             traceback.print_exc()
             resp = INTERNAL_ERROR
         return resp
 
 
-# PCIe Switches API
-class PCIeSwitchesAPI(Resource):
+# EgResource Collection API
+class EgResourceCollectionAPI(Resource):
     def __init__(self):
         self.rb = g.rest_base
         self.config = {
-            '@odata.context': self.rb + '$metadata#PCIeSwitches',
-            '@odata.id': self.rb + 'PCIeSwitches',
-            '@odata.type': '#PCIeSwitches.1.0.0.PCIeSwitches',
-            'Name': 'PCIe Switches Collection',
+            '@odata.context': self.rb + '$metadata#EgResourceCollection',
+            '@odata.id': self.rb + 'EgResourceCollection',
+            '@odata.type': '#EgResourceCollection.1.0.0.EgResourceCollection',
+            'Name': 'EgResource Collection',
             'Links': {}
         }
         self.config['Links']['Member@odata.count'] = len(member_ids)
@@ -128,8 +128,6 @@ class PCIeSwitchesAPI(Resource):
     def get(self):
         try:
             resp = self.config, 200
-        except OSError:
-            resp = error_response('Resources directory does not exist', 400)
         except Exception:
             traceback.print_exc()
             resp = INTERNAL_ERROR
@@ -139,11 +137,31 @@ class PCIeSwitchesAPI(Resource):
     # Todo - Fix so the config can be passed in the data.
     def post(self):
         try:
-            g.api.add_resource(PCIeSwitchAPI, '/redfish/v1/PCIeSwitches/<string:ident>')
+            g.api.add_resource(EgResourceAPI, '/redfish/v1/EgResources/<string:ident>')
             resp=self.config,200
-        except PathError:
-            resp = error_response('Attribute Does Not Exist', 404)
         except Exception:
             traceback.print_exc()
             resp = INTERNAL_ERROR
+        return resp
+
+
+# Used to create a resource instance internally
+class CreateEgResource(object):
+    def __init__(self):
+        print ('CreateEgResource init called')
+
+    def put(self,ident):
+        print ('CreateEgResource put called')
+        try:
+            global config
+            config=get_EgResource_template(g.rest_base,ident)
+            members.append(config)
+            member_ids.append({'@odata.id': config['@odata.id']})
+            # Add APIs for subordinate resourcs
+            # g.api.add_resource(SubordinateResAPI, '/redfish/v1/EgResources/<string:ch_id>/SubordinateRes')
+            resp = config, 200
+        except Exception:
+            traceback.print_exc()
+            resp = INTERNAL_ERROR
+        print ('CreateEgResource put exit')
         return resp