+++ api_emulator/redfish/pcie_port_api.py
@@ -9,7 +9,7 @@ import g
 
 # Flask Imports
 from flask import Flask, request, make_response, render_template
-from flask.ext.restful import reqparse, Api, Resource
+from flask_restful import reqparse, Api, Resource
 
 from .templates.pcie_port import get_PCIePort_template