+++ emulator.py
@@ -15,7 +15,7 @@ import xml.etree.ElementTree as ET
 
 # Flask Imports
 from flask import Flask, request, make_response, render_template
-from flask.ext.restful import reqparse, Api, Resource
+from flask_restful import reqparse, Api, Resource
 
 import g