+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # Change Log
 
-## [0.9.0] - 2016-09-22
+## 2016-09-22 [0.9.0]
 - Initial Release
-
+## 2016-03-17
+- Fixes to run with Python 3.5
+## 2017-05-21
+- Change references from flask.ext.restful to flask_restful, since package is obsolete
+- Add example files, eg_resource_api.py and eg_resource.py, showing how dynamic objects are coded.